This commit is about enabling :select and :range support coming from domain IR in support of ticket 30. Blackscholes still seems to work and the code from ticket30 now looks good in terms of the lowered parfors that are created and there is now a problem in cgen.
With :select and :range support, not all parfor nodes are simply indexed so adding a flag in parfor to remember which ones are simply indexed and which ones are not.
Adding data structure to hold range information.
Adding data structure InputInfo that is used by mmap and mmap! now to record information associated with each input.  That information includes:
   1) The array name.
   2) An array of RangeData information for each dimension of the array.
   3) An array of SymbolNode that holds the offset difference between the iteration space and the range space.
   4) elementTemp holds the SymbolNode into which the element for this point of the iteration space is stored for this range.
   5) pre_offsets holds the statements that need to be inserted into the parfor's prestatements to load the correct values into the range_offset variables (#3).
Adding a function rangeSize to compute the size of the range space (last - start + 1).  This is probably buggy because we don't deal with non-default skip values.
Changing lowering for mmap and mmap! so that when a range is input that you get the iteration space length from this range (through rangeSize()) rather than arraysize of the containing array.
Do the index calculation when reading a ranged array so that we add the range offset computed in the pre-statements to the corresponding parfor index symbol.