Changing parallel IR to generate and reconstruct jl_alloc_array_Nd calls like Julia 0.4 is now generating them. The primary difference is that a SimpleVector is used instead of a Tuple for the third argument that specifies the types of the arguments to the ccall.
Add some support for 3D array construction.

Estimate instruction count sometime start in a body and needs to be given a LambdaInfo and sometimes it starts with a :lambda Expr and so needs a
"nothing" LambdaInfo to start.  Support for the latter case has been added.

Support in pir_live_cb has been added to convert non-standard unsafe_arrayref and usafe_arrayset into their standard versions since it is only
for the purpose of liveness analysis which is the same between the two functions.