Make it easy to embed ParallelAccelerator into a Julia distribution.
Adding embed() and embed(path) functions that try to find a Julia source distribution,
create a userimg.jl in base, and re-make Julia.

The version with no paramters uses the Julia global "JULIA_HOME" to try to find the
source distribution and then calls the version that takes an explicit path.