Importing Base.hash and Base.isequal so we can get Dictionaries working with user-defined types. Overload hash and isequal for the following types so that our range correlation dictionary or arrays of DimensionSelector will work correctly:    1) Rangedata    2) RangeExprs    3) MaskSelector    4) SingularSelector    5) Array{DimensionSelector,1}    6) SymbolNode
Switch PIRParForAst to remember an InputInfo from its first array for later fusion correlations.
Create a range_correlation Union type DimensionSelector for the various possible types of ranges and use that instead of generic Any in range_correlation.

Add code to detect if all ranges are masks in which case we can associate the left-hand side with the right-hand side array to which the masks are being applied.

Add routine print_correlations to print array length, symbol, and range correlations but only if they are non-empty.