With this update, vector-style centroid update in kmeans now works and all tests pass including the reduction tests.
In the reduction lowering code, changed the loop to go over num_dim_inputs instead of the number of dimensions of the incoming array
because of singular dimension selectors at the end of the indexing expression.

Made the SingularSelector code for non-trailing singular selections to behave more like the other selection types.  In particular,
it sets a save_array_lenth in the pre-statements, sets input_array_rangeconds to nothing and it now falls through to the
existing code to add this dimension to the loopNests.