Fixes #issue71.
Move function isSymbolsUsed to fusion file since it is only used there.

I guess previously the first param to isSymbolsUsed was an array or a set but then at some point it got changed to a Dict.
When this happened, the iteration through the collection stopped being over SymGen but over Pair{SymGen,SymGen}.  This latter
type would fail to match against Symbols in the second parfor's use or def sets.  The result is that reduction variables
computed in the first parfor and used by the second parfor would not be detected.  Fix to compare both parts of the pair
to the use and def sets.