Pass linfo to some additional function like from_header, from_UDTs, from_decl.
Adding routines to check if something is in the symbol table, lookup something from the symbol table and
change a type in the symbol table.
For 0.5, we convert SlotNumber to Symbol before doing the lookup.
These routines aren't being used yet as I'm searching for a bug with some big changes I've done.

fromRHSVar uses string(typeof(ast)) instead of "GenSym" to accommodate 0.5 typealias of SSAValue to GenSym.

Use from_lambda in from_root_nonentry instead of generic from_expr.