ParallelIR detect if correlation size constant in multi assignment
Previous method of detecting multiple assignments separately and returning negative correlation
cannot handle the case where all the sizes are the same in the assignments.
Handling multiple assignment case is move to equivalence class pass
so it has the size information.
Kmeans example needs this since there are multiple assignments to centroids
but they are of the same size.