Adding getAllAliases that takes a LHSVar that gets expanded to a Set and call the existing alias routine.
Adding is_eliminated_arraysize that takes a set of allocations that were removed from the previous parfor's prestatements during fusion
along with a set of aliases and looks at each statement in the next parfor's prestatements to see if there is an arraysize expression
for an eliminated intermediate array.  If so then eliminate the arraysize expression as well.

Extend hasNoSideEffects to objects of type DataType.

Adding matching parenthesis in nested_function_exprs to make it easy to go from beginning to end of that routine's output
or vice versa.