Simplifying mk_arraylen_expr to just use "x" instead of ":($x)" and also to deepcopy x to prevent aliasing issues.
Deepcopy a copule other places where things get put into the ast to prevent aliasing issues.

Improve parfor liveness summarization to make the use set equal to the intersection of all the symbols used in the parfor unioned with those live_in to the parfor.  If they aren't live in then they must be used after def in the parfor.  If they are live_in but not used then it is wrong to say the parfor uses them.