Add overloaded show method for TaskInfo.
Remove addition of LineNumberNode.

Remove some old, commented out lines.

Fix a problem with reduction where the parfor had replaceExprWithDict run on it and that mapped the reductions to the
new task function's namespace but that same reduction information was used in the calling function to create the
reduction across what the threads returned.  The solution was to copy the original calling namespace reduction information
before running replaceExprWithDict on the parfor.

Remove some old code that created C reduction functions.